Refactor: Modularize visualize_current_and_field_history
